🔖 fix(1701): Correct typo in problem description
The changes in this commit fix a typo in the problem description for the "Average Waiting Time" problem (1701). The typo is corrected from "Solution_1717" to "Solution" to accurately describe the expected output format.
