+++ solutions/1518_Water_Bottles/Solution_1518.java
@@ -1,4 +1,4 @@
-class Solution {
+public class Solution_1518 {
     public int numWaterBottles(int numBottles, int numExchange) {
 
         return numBottles + (numBottles-1)/(numExchange-1);
