+++ solutions/1717_Maximum_Score_From_Removing_Substrings/Solution_1717.java
@@ -1,4 +1,4 @@
-class Solutions {
+public class Solution_1717 {
     public int maximumGain(String s, int x, int y) {
         int res = 0;
         String top, bot;
