--- solutions/2058_Find_the_Minimum_and_Maximum_Number/Solutions.java
@@ -1,82 +1,0 @@
-
-class Solutions {
-
-    public static void main(String[] args) {
-
-    }
-    public int[] nodesBetweenCriticalPoints(ListNode head) {
-
-        class Solution {
-
-            public int[] nodesBetweenCriticalPoints(ListNode head) {
-                int[] result = { -1, -1 };
-
-                // Initialize minimum distance to the maximum possible value
-                int minDistance = Integer.MAX_VALUE;
-
-                // Pointers to track the previous node, current node, and indices
-                ListNode previousNode = head;
-                ListNode currentNode = head.next;
-                int currentIndex = 1;
-                int previousCriticalIndex = 0;
-                int firstCriticalIndex = 0;
-
-                while (currentNode.next != null) {
-                    // Check if the current node is a local maxima or minima
-                    if (
-                            (currentNode.val < previousNode.val &&
-                                    currentNode.val < currentNode.next.val) ||
-                                    (currentNode.val > previousNode.val &&
-                                            currentNode.val > currentNode.next.val)
-                    ) {
-                        // If this is the first critical point found
-                        if (previousCriticalIndex == 0) {
-                            previousCriticalIndex = currentIndex;
-                            firstCriticalIndex = currentIndex;
-                        } else {
-                            // Calculate the minimum distance between critical points
-                            minDistance = Math.min(
-                                    minDistance,
-                                    currentIndex - previousCriticalIndex
-                            );
-                            previousCriticalIndex = currentIndex;
-                        }
-                    }
-
-                    // Move to the next node and update indices
-                    currentIndex++;
-                    previousNode = currentNode;
-                    currentNode = currentNode.next;
-                }
-
-                // If at least two critical points were found
-                if (minDistance != Integer.MAX_VALUE) {
-                    int maxDistance = previousCriticalIndex - firstCriticalIndex;
-                    result = new int[] { minDistance, maxDistance };
-                }
-
-                return result;
-            }
-        }
-
-    }
-
-
-    class ListNode {
-        int val;
-        ListNode next;
-
-        ListNode() {
-        }
-
-        ListNode(int val) {
-            this.val = val;
-        }
-
-        ListNode(int val, ListNode next) {
-            this.val = val;
-            this.next = next;
-        }
-    }
-
-}
