+++ solutions/2191_Sort_the_Jumbled_Numbers/Main.java
@@ -1,0 +1,26 @@
+public class Main {
+    public static void main(String[] args) {
+        // Create an instance of the Solution class
+        Solution_2191 solution = new Solution_2191();
+
+        // Test case
+        int[] mapping = {8, 9, 4, 0, 2, 1, 3, 5, 7, 6};
+        int[] nums = {991, 338, 38};
+
+        // Call the sortJumbled method
+        int[] result = solution.sortJumbled(mapping, nums);
+
+        // Print the result
+        System.out.print("Sorted Jumbled Numbers: ");
+        for (int num : result) {
+            System.out.print(num + " ");
+        }
+    }
+}
+
+
+class Solution_2191 {
+    public int[] sortJumbled(int[] mapping, int[] nums) {
+        return new int[0];
+    }
+}
